+++ R/manyenviron-references.R
@@ -1,14 +1,34 @@
-#' references database documentation
+#' references datacube documentation
 #'
-#' @format The references database is a list that contains the
+#' @format The references datacube is a list that contains the
 #' following 1 datasets: ECOLEX_REF.
 #' For more information and references to each of the datasets used,
-#' please use the `data_source()` and `data_contrast()` functions.
+#' please use the `manydata::call_sources()` and `manydata::compare_dimensions()` functions.
 #'\describe{
 #' \item{ECOLEX_REF: }{A dataset with 1164 observations and the following
 #' 3 variables: treatyID1, treatyID2, RefType.}
 #' }
-#'
+#' @source
+#' \itemize{
+#' \item{ECOLEX_REF: }{
+#' R. E. Kim and B. Mackey. “International environmental law as a complex adaptive system”. In: _International Environmental Agreements: Politics, Law and Economics_ 14.1 (2014), pp. 5-24.}
+#' }
+#' @section URL:
+#' \itemize{
+#' \item{ECOLEX_REF: }{NA}
+#' }
+#' @section Mapping:
+#' \itemize{
+#' \item{ECOLEX_REF: }{
+#' Variable Mapping
+#' 
+#' |  *from*  | *to*
+#' |:------------:|:------------:|
+#' | NA | NA |
+#' 
+#' }
+#' }
+#' @md
 #' @details
 #' ``` {r, echo = FALSE, warning = FALSE}
 #' lapply(references, messydates::mreport)