+++ data-raw/agreements/HEIDI/prepare-HEIDI.R
@@ -1,27 +1,30 @@
 # HEIDI Preparation Script
 
 # This is a template for importing, cleaning, and exporting data
-# ready for many packages universe.
+# for the 'many' packages.
 
 # Stage one: Collecting data
 HEIDI <- readxl::read_excel("data-raw/agreements/HEIDI/heidi_dataset.xlsx")
-HEIDI$signature.date <- openxlsx::convertToDate(HEIDI1$signature.date)
+HEIDI$signature.date <- ifelse(stringr::str_detect(HEIDI$signature.date,
+                                                   "^([:digit:]{2})-([:digit:]{2})-([:digit:]{4})$|^([:digit:]{4})-([:digit:]{2})-([:digit:]{2})$"),
+                               as.Date(HEIDI$signature.date),
+                               openxlsx::convertToDate(HEIDI$signature.date))
+HEIDI$signature.date <- as.Date(HEIDI$signature.date, origin = "1970-01-01")
 
 # Stage two: Correcting data
 # In this stage you will want to correct the variable names and
 # formats of the 'HEIDI' object until the object created
 # below (in stage three) passes all the tests.
 HEIDI <- as_tibble(HEIDI) %>%
-  manydata::transmutate(Title = manypkgs::standardise_titles(`Name.of.the.agreement`,
-                                                             api_key = api),
+  manydata::transmutate(Title = manypkgs::standardise_titles(`Name.of.the.agreement`),
                         Signature = messydates::as_messydate(`signature.date`)) %>%
-  dplyr::mutate(Beg = Signature) %>%
+  dplyr::mutate(Begin = Signature) %>%
   dplyr::rename(heidiID = ID) %>%
-  dplyr::select(heidiID, Title, Beg, Signature) %>%
-  dplyr::arrange(Beg)
+  dplyr::select(heidiID, Title, Begin, Signature) %>%
+  dplyr::arrange(Begin)
 
 # Add treaty_ID column
-HEIDI$treatyID <- manypkgs::code_agreements(HEIDI, HEIDI$Title, HEIDI$Beg)
+HEIDI$treatyID <- manypkgs::code_agreements(HEIDI, HEIDI$Title, HEIDI$Begin)
 
 # Add Lineage column
 HEIDI$Lineage <- manypkgs::code_lineage(HEIDI$Title)
@@ -32,9 +35,9 @@ HEIDI <- dplyr::left_join(HEIDI, manyID, by = "treatyID")
 
 # Re-order the columns
 HEIDI <- HEIDI %>%
-  dplyr::select(manyID, Title, Beg, Signature,
+  dplyr::select(manyID, Title, Begin, Signature,
                 Lineage, treatyID, heidiID) %>%
-  dplyr::arrange(Beg)
+  dplyr::arrange(Begin)
 
 # manypkgs includes several functions that should help cleaning
 # and standardising your data.
@@ -43,7 +46,7 @@ HEIDI <- HEIDI %>%
 # Stage three: Connecting data
 # Next run the following line to make IEADB available
 # within the package.
-manypkgs::export_data(HEIDI, database = "agreements",
+manypkgs::export_data(HEIDI, datacube = "agreements",
                       URL = "https://www.chaire-epi.ulaval.ca/en/data/heidi")
 # This function also does two additional things.
 # First, it creates a set of tests for this object to ensure