+++ data-raw/agreements/HUGGO/prepare-HUGGO.R
@@ -1,7 +1,7 @@
 # HUGGO Preparation Script
 
 # This is a template for importing, cleaning, and exporting data
-# ready for the many packages universe.
+# for the 'many' packages.
 
 # Stage one: Collecting data
 HUGGO1 <- readr::read_csv("data-raw/agreements/HUGGO/EnvGov Nodes-Table 1 VERS2.csv")
@@ -121,199 +121,15 @@ manyID <- manypkgs::condense_agreements(manyenviron::agreements,
 
 HUGGO <- dplyr::left_join(HUGGO, manyID, by = "treatyID")
 
-# Extract treaty texts
-# Create a texs "database" to apply consolidate()
-texts <- list(HUGGO, IEADB, ECOLEX, CIESIN, HEIDI)
-texts <- manydata::consolidate(texts, "any", "any",
-                               resolve = "coalesce", key = "manyID")
-
-# Step two: extract treaty texts from IEADB website
-# This requires the original IEADB dataset with the ieadbID column
-IEADB_original <- readr::read_csv("data-raw/agreements/IEADB/treaties.csv")
-IEADB_original <- dplyr::as_tibble(IEADB_original) %>%
-  dplyr::rename(ieadbID = `IEA# (click for add'l info)`,
-                TreatyText = `Treaty Text`) %>%
-  dplyr::select(ieadbID, TreatyText) %>%
-  dplyr::filter(TreatyText == "Treaty Text**") %>%
-  dplyr::arrange(ieadbID)
-IEADB_original$ieadbID <- as.character(IEADB_original$ieadbID)
-
-# Then, rvest package is used to web scrap IEADB treaty texts pages
-base <- "https://iea.uoregon.edu/treaty-text/"
-IEADB_original$TreatyText <- lapply(IEADB_original$ieadbID,
-                                    function(s) tryCatch(rvest::read_html(paste0(base, s)) %>%
-                                                           rvest::html_nodes(".even") %>%
-                                                           rvest::html_text(),
-                                                         error = function(e) {
-                                                           as.character("Not found")
-                                                         } %>%
-                                                           paste(collapse = ",")))
-# A source column is also added
-IEADB_original$Source <- lapply(IEADB_original$ieadbID,
-                                function(s) tryCatch(rvest::read_html(paste0(base, s)) %>%
-                                                       rvest::html_nodes(".views-field-views-conditional") %>%
-                                                       rvest::html_text(),
-                                                     error = function(e) {
-                                                       as.character("Not found")
-                                                     } %>%
-                                                       paste(collapse = ",")))
-
-# Step three: join IEADB text column with the consolidated version of
-# manyenviron
-texts <- dplyr::left_join(texts,
-                          IEADB_original,
-                          by = "ieadbID")
-texts <- as_tibble(texts) %>%
-  dplyr::select(treatyID, ieadbID, gengID, ecolexID,
-                manyID, Title, Beg, TreatyText, Source, url)
-
-# Step four: complement the dataset with ECOLEX treaty texts (TO BE IMPROVED)
-# Filter observations with text = NULL and with an ecolexID
-ecolex_text <- texts %>%
-  dplyr::filter(!is.na(ecolexID) & TreatyText == "NULL")
-
-ecolex_text$ecolexID2 <- stringr::str_remove_all(ecolex_text$ecolexID, "-")
-
-# Use pdftools package to extract treaty texts
-base <- "http://www.ecolex.org/server2neu.php/libcat/docs/TRE/Full/En/"
-ecolex_text$TreatyText <- lapply(ecolex_text$ecolexID,
-                                 function(s) tryCatch(pdftools::pdf_text(paste0(base, s, ".pdf")),
-                                                      error = function(e) {
-                                                        as.character("Not found")
-                                                      }))
-ecolex_text$TreatyTextb <- lapply(ecolex_text$ecolexID2,
-                                  function(s) tryCatch(pdftools::pdf_text(paste0(base, s, ".pdf")),
-                                                       error = function(e) {
-                                                         as.character("Not found")
-                                                       }))
-
-
-base2 <- "http://www.ecolex.org/server2neu.php/libcat/docs/TRE/Full/Other/"
-ecolex_text$TreatyText2 <- lapply(ecolex_text$ecolexID,
-                                  function(s) tryCatch(pdftools::pdf_text(paste0(base2, s, ".pdf")),
-                                                       error = function(e) {
-                                                         as.character("Not found")
-                                                       }))
-ecolex_text$TreatyText2b <- lapply(ecolex_text$ecolexID2,
-                                   function(s) tryCatch(pdftools::pdf_text(paste0(base2, s, ".pdf")),
-                                                        error = function(e) {
-                                                          as.character("Not found")
-                                                        }))
-
-# base3 = "http://www.ecolex.org/server2neu.php/libcat/docs/TRE/Full/Fr/"
-# ecolex_text$TreatyText3 <- lapply(ecolex_text$ecolexID,
-#                                   function(s) tryCatch(pdftools::pdf_text(paste0(base3, s, ".pdf")),
-#                                                        error = function(e){as.character("Not found")}))
-# ecolex_text$TreatyText3b <- lapply(ecolex_text$ecolexID2,
-#                                    function(s) tryCatch(pdftools::pdf_text(paste0(base3, s, ".pdf")),
-#                                                         error = function(e){as.character("Not found")}))
-#
-# base4 = "http://www.ecolex.org/server2neu.php/libcat/docs/TRE/Full/Sp/"
-# ecolex_text$TreatyText4 <- lapply(ecolex_text$ecolexID,
-#                                   function(s) tryCatch(pdftools::pdf_text(paste0(base4, s, ".pdf")),
-#                                                        error = function(e){as.character("Not found")}))
-# ecolex_text$TreatyText4b <- lapply(ecolex_text$ecolexID2,
-#                                    function(s) tryCatch(pdftools::pdf_text(paste0(base4, s, ".pdf")),
-#                                                         error = function(e){as.character("Not found")}))
-
-ecolex_text$TreatyText <- dplyr::na_if(ecolex_text$TreatyText, "Not found")
-ecolex_text$TreatyTextb <- dplyr::na_if(ecolex_text$TreatyTextb, "Not found")
-ecolex_text$TreatyText2 <- dplyr::na_if(ecolex_text$TreatyText2, "Not found")
-ecolex_text$TreatyText2b <- dplyr::na_if(ecolex_text$TreatyText2b, "Not found")
-
-ecolex_text$Text <- dplyr::coalesce(ecolex_text$TreatyText,
-                                    ecolex_text$TreatyTextb,
-                                    ecolex_text$TreatyText2,
-                                    ecolex_text$TreatyText2b)
-
-ecolex_text$url <- ifelse(!is.na(ecolex_text$TreatyText), "http://www.ecolex.org/server2neu.php/libcat/docs/TRE/Full/En/",
-                                    ifelse(!is.na(ecolex_text$TreatyTextb), "http://www.ecolex.org/server2neu.php/libcat/docs/TRE/Full/En/",
-                                           ifelse(!is.na(ecolex_text$TreatyText2), "http://www.ecolex.org/server2neu.php/libcat/docs/TRE/Full/Other/",
-                                                  ifelse(!is.na(ecolex_text$TreatyText2b), "http://www.ecolex.org/server2neu.php/libcat/docs/TRE/Full/Other/",
-                                                         NA))))
-
-ecolex_text <- ecolex_text %>%
-  dplyr::filter(!is.na(Text)) %>%
-  dplyr::select(ecolexID, Title, Beg, Text, url) %>%
-  dplyr::rename(TreatyText = Text)
-
-# Step five: join ECOLEX texts to the agreements dataset
-texts <- dplyr::left_join(texts,  ecolex_text, by = c("ecolexID", "Title",
-                                                      "Beg","TreatyText",
-                                                      "url"))
-
-# Step six: Clean texts
-# manypkgs includes several functions that should help cleaning
-# and standardizing your data.
-# Please see the vignettes or website for more details.
-texts <- texts %>%
-  dplyr::mutate(TreatyText = manypkgs::standardise_treaty_text(TreatyText))
-
-texts <- dplyr::as_tibble(texts) %>%
-  dplyr::select(manyID, Title, Beg, TreatyText, url, Source)
-
-# Bind data
-HUGGO <- dplyr::left_join(HUGGO, texts, by = c("manyID", "Title", "Beg", "url"))
-
-# reorder variables
-HUGGO <- dplyr::relocate(HUGGO, manyID, Title, Beg, End, Signature,
-                         Force, AgreementType, DocType, GeogArea,
-                         gengID, ieaID, ecolexID, treatyID)
-
-# make sure all vars are correctly coded as NA if necessary
-HUGGO <- HUGGO %>%
-  dplyr::mutate(across(everything(), ~stringr::str_replace_all(., "^NA$",
-                                                               NA_character_))) %>%
-  dplyr::distinct() %>%
-  mutate(Signature = messydates::as_messydate(Signature),
-         Force = messydates::as_messydate(Force),
-         Beg = messydates::as_messydate(Beg),
-         End = messydates::as_messydate(End)) %>%
-  dplyr::distinct() %>%
-  dplyr::arrange(Beg)
-
-# Standardising treaty texts
-HUGGO <- HUGGO %>%
-  dplyr::mutate(MainText = ifelse(!grepl("APPENDIX | ANNEX", unlist(TreatyText), perl = T),
-                                  TreatyText,
-                                  stringr::str_remove(HUGGO$TreatyText,
-                                                      "APPENDIX.* | ANNEX.*")))
-HUGGO <- HUGGO %>%
-  dplyr::mutate(AppendixText = ifelse(grepl("APPENDIX", unlist(TreatyText), perl = T),
-                                      stringr::str_extract(HUGGO$TreatyText,
-                                                           "APPENDIX.*"),
-                                      NA))
-HUGGO <- HUGGO %>%
-  dplyr::mutate(AnnexText = ifelse(grepl("ANNEX", unlist(TreatyText), perl = T),
-                                   stringr::str_extract(HUGGO$TreatyText,
-                                                        "ANNEX.*"),
-                                   NA))
-
-# Checked_HUGGO and Confirmed_HUGGO variables to
-# track progress on manually correcting entries
-# Checked_HUGGO: code 1 when the entire row's
-# observations have been verified and updated
-# Confirmed_HUGGO: list variables for which the
-# observation could be verified and confirmed.
-# Eg. List 'Signature' in `Confirmed_HUGGO` if
-# the Signature date was found and verified in
-# the treaty text or in a manual online search.
-HUGGO$Checked_HUGGO <- NA
-HUGGO$Confirmed_HUGGO <- NA
-
-# Add Changes var to log changes that are manually added for each agreement.
-HUGGO$Changes <- NA
-
 # Stage three: Merging verified data and additional treaties into HUGGO dataset
 
-# Step one: Merge data frames with verified metadata of treaties within and
-# without HUGGO
+# Step one: Merge original HUGGO data with verified metadata of treaties in the agreements datacube
 
 # Load original HUGGO
 HUGGO_or <- manyenviron::agreements$HUGGO
 
 # Load data frame with verified metadata (of treaties present in HUGGO)
-HUGGO7 <- read.csv("data-raw/agreements/HUGGO/HUGGO_verified.csv")
+HUGGO7 <- read.csv("data-raw/agreements/HUGGO/HUGGO_reconciled.csv")
 
 # Load data frame with verified metadata (of treaties not present in HUGGO)
 HUGGO8 <- read.csv("data-raw/agreements/HUGGO/HUGGO_additional.csv")
@@ -442,7 +258,7 @@ HUGGO_new <- HUGGO_new %>%
                   Data, Coded, Abstract) %>%
   dplyr::arrange(Beg)
 
-# Step two: Clean duplicate rows
+# Step two: Remove duplicate rows
 
 # First row NA remove
 HUGGO_new <- HUGGO_new[-(which(is.na(HUGGO_new$manyID))), ]
@@ -839,7 +655,7 @@ HUGGO_new <- HUGGO_new[-(which(grepl(title, HUGGO_new$Title) &
                                  is.na(HUGGO_new$AgreementType) &
                                  HUGGO_new$Beg == "2001-10-16")), ]
 
-# Step three: TreatyText and Language columns
+# Step three: Add TreatyText and Language columns
 
 # Add columns to show if the raw text of the agreements has been collected and
 # to indicate its language
@@ -848,19 +664,127 @@ HUGGO_new <- HUGGO_new %>%
   dplyr::mutate(TreatyText = NA) %>%
   dplyr::mutate(Language = NA)
 
-# Step four: Standardise date columns, arrange by Beg,
+# Step four: Standardise date columns, arrange by Begin,
 # and push HUGGO_new to HUGGO
 HUGGO <- HUGGO_new %>%
-  dplyr::mutate(Beg = messydates::as_messydate(Beg),
+  dplyr::mutate(Begin = messydates::as_messydate(Beg),
                 Signature = messydates::as_messydate(Signature),
                 Force = messydates::as_messydate(Force),
                 End = messydates::as_messydate(End)) %>%
-  dplyr::arrange(Beg)
-          
-# Stage four: Connecting data
+  dplyr::arrange(Begin)
+
+# Stage four: Matching agreements with titles in different languages
+noneng <- HUGGO %>%
+  # subset agreements with non-English titles but match a row with an English title
+  dplyr::filter(manyID == "UB08IB_1893A" | manyID == "ESP-FRA[DGP]_1900A" |
+                  manyID == "ED19DL_1905A" | manyID == "CHE-FRA[DDK]_1930A" |
+                  manyID == "ARG-PRY[DRP]_1939A" | manyID == "CP09RP_1971A" |
+                  manyID == "CP09RP_1971A" | manyID == "TRTDDY_1973A" |
+                  manyID == "AS13DH_1978A" | manyID == "AT24JD_1979A" |
+                  manyID == "AD25ED_1979A" | manyID == "MAR-ROU[SDR]_1979A" |
+                  manyID == "BEL-LUX[EDL]_1980A" | manyID == "COL-ECU[SPA]_1994A" |
+                  manyID == "FRA-NOR[PLM]_1995A" | manyID == "AS32AS_1995A" |
+                  manyID == "AM17LS_1996A" | manyID == "ARG-PRY[LPN]_1996A" |
+                  manyID == "AE21LM_1997A" | manyID == "AM17LD_1997A" |
+                  manyID == "ARG-BOL[RDB]_1998A" | manyID == "TD13DP_1998A" |
+                  manyID == "AP17RT_2002A" | manyID == "CHN-TUN[DTM]_2002A" |
+                  manyID == "DZA-TUR[RDT]_2002A" | manyID == "AM23AE_2002A" |
+                  manyID == "AE38LD_2002A" | manyID == "AE19DL_2003A" |
+                  manyID == "AE24LJ_2003A"| manyID == "EL50LJ_2004A" |
+                  manyID == "ARG-BOL[RDB]_2004P" | manyID == "AB22LA_2005A" |
+                  manyID == "HND-MEX[EUM]_2005A" | manyID == "CHL-DZA[RDC]_2005A" |
+                  manyID == "DZA-PER[RDP]_2005A" | manyID == "AC46LJ_2005A" |
+                  manyID == "CD34RM_2005N1" | manyID == "MD19CC_2005S" |
+                  manyID == "AM15DC_2006A" | manyID == "AC43LO_2006A" |
+                  manyID == "NIC-VEN[LRD]_2007A" | manyID == "AC33LJ_2007A"|
+                  manyID == "ARG-ECU[LHE]_2007A" | manyID == "AC11PB_2007A" |
+                  manyID == "AD33LN_2007A" | manyID == "AC34LD_2007A" |
+                  manyID == "AD25LF_2008A" | manyID == "BRA-CRI[DSL]_2008A"|
+                  manyID == "BRA-CRI[PDB]_2008A" | manyID == "BRA-CRI[DCE]_2008P" |
+                  manyID == "AE22LE_2008A" | manyID == "AS35LO_2008A" |
+                  manyID == "CP09LA_2012O"| manyID == "CC09DM_1954A" |
+                  manyID == "VB13GV_2008O" | manyID == "CS18PS_1954A"|
+                  manyID == "CS18PS_1954A")
+noneng <- noneng %>%
+  dplyr::rename(match = manyID, Orig_noneng_title = Title) %>%
+  dplyr::arrange(match) %>%
+  # add in matching list of manyIDs of agreements with English titles
+  dplyr::mutate(manyID = c("FRA-URY[CCM]_2005A", "SI04DB_2007P", "FRA-USA[PPP]_2007A",
+                           "UKR-ALG[PPA]_2007A", "FRA-TUN[CCT]_2006P:UNFCCC_1992A",
+                           "BRA-FRA[MKP]_2005P", "ESP-SEN[FAD]_1979A", "FRA-ZAF[ECT]_2008A",
+                           "CHN-FRA[NEB]_2007A", "FRA-GAB[HWD]_2003A", "FRA-TUR[FEP]_1997A",
+                           "VEN-ZAF[ENR]_2008A", "GF13JT_2003A", "FRA-KIR[EZT]_2002A",
+                           "COG-ALG[MRD]_2006A", "FRA-LVA[MFR]_1997A", "FRA-TUR[MFP]_1996A",
+                           "ARG-ALG[DSI]_2002A", "TUN-ALG[MBD]_2002A", "ARG-BOL[EIB]_1998A",
+                           "ARG-BOL[ENB]_2004P", "ARG-ECU[SHE]_2007A", "ARG-PRY[LRP]_1939A",
+                           "ARG-PRY[PNP]_1996A", "COL-HTI[MSA]_1978A", "CHE-FRA[LGP]_1995A",
+                           "DEU-FRA[SFP]_2008A",
+                           "ESP-MAR[MFR]_1979A","KB04WS_1980A","BRA-CRI[MCT]_2008P",
+                           "BRA-CRI[ASL]_2008P","BRA-CRI[ABP]_2008P","CDSMZM_1954P20",
+                           "ESP-PRT[ISM]_2005N","CHE-FRA[BDK]_1930A","CHL-ALG[EMD]_2005A",
+                           "CHN-TUN[PMT]_2002A","COL-ECU[ARF]_1994A","ESTANA_2012A",
+                           "STDYPR_1971A", "INVEPR_1971A", "ES06SP_1954A","PER-ALG[EMD]_2005A",
+                           "TUR-ALG[END]_2002A","CF08CL_1905N11","FRA-MCO[TMS]_2004A",
+                           "FS10PJ_1900A","FRA-NOR[SSP]_1995A","HND-MEX[MDM]_2005A",
+                           "MAR-ROU[TMS]_1979A","ESP-GTM[ECC]_2005S","NIC-VEN[ENS]_2007A",
+                           "ECU-PER[CMG]_1998A","YCYRTT_1973A","AH12IF_1893A", "TTICPO_2008A"
+  ))
+noneng <- dplyr::select(noneng, manyID, Orig_noneng_title, Begin, match)
+# Non-English titles of the same agreement with an English title added in variable 'Orig_noneng_title'
+# manyID of the non-English agreement added in variable 'match'
+HUGGO <- dplyr::left_join(HUGGO, noneng, by = c("manyID", "Begin"))
+# remove rows with non-English titles identified above
+HUGGO <- HUGGO[-which(HUGGO$manyID == "UB08IB_1893A" | HUGGO$manyID == "ESP-FRA[DGP]_1900A" |
+                         HUGGO$manyID == "ED19DL_1905A" | HUGGO$manyID == "CHE-FRA[DDK]_1930A" |
+                         HUGGO$manyID == "ARG-PRY[DRP]_1939A" | HUGGO$manyID == "CP09RP_1971A" |
+                         HUGGO$manyID == "CP09RP_1971A" | HUGGO$manyID == "TRTDDY_1973A" |
+                         HUGGO$manyID == "AS13DH_1978A" | HUGGO$manyID == "AT24JD_1979A" |
+                         HUGGO$manyID == "AD25ED_1979A" | HUGGO$manyID == "MAR-ROU[SDR]_1979A" |
+                         HUGGO$manyID == "BEL-LUX[EDL]_1980A" | HUGGO$manyID == "COL-ECU[SPA]_1994A" |
+                         HUGGO$manyID == "FRA-NOR[PLM]_1995A" | HUGGO$manyID == "AS32AS_1995A" |
+                         HUGGO$manyID == "AM17LS_1996A" | HUGGO$manyID == "ARG-PRY[LPN]_1996A" |
+                         HUGGO$manyID == "AE21LM_1997A" | HUGGO$manyID == "AM17LD_1997A" |
+                         HUGGO$manyID == "ARG-BOL[RDB]_1998A" | HUGGO$manyID == "TD13DP_1998A" |
+                         HUGGO$manyID == "AP17RT_2002A" | HUGGO$manyID == "CHN-TUN[DTM]_2002A" |
+                         HUGGO$manyID == "DZA-TUR[RDT]_2002A" | HUGGO$manyID == "AM23AE_2002A" |
+                         HUGGO$manyID == "AE38LD_2002A" | HUGGO$manyID == "AE19DL_2003A" |
+                         HUGGO$manyID == "AE24LJ_2003A"| HUGGO$manyID == "EL50LJ_2004A" |
+                         HUGGO$manyID == "ARG-BOL[RDB]_2004P" | HUGGO$manyID == "AB22LA_2005A" |
+                         HUGGO$manyID == "HND-MEX[EUM]_2005A" | HUGGO$manyID == "CHL-DZA[RDC]_2005A" |
+                         HUGGO$manyID == "DZA-PER[RDP]_2005A" | HUGGO$manyID == "AC46LJ_2005A" |
+                         HUGGO$manyID == "CD34RM_2005N1" | HUGGO$manyID == "MD19CC_2005S" |
+                         HUGGO$manyID == "AM15DC_2006A" | HUGGO$manyID == "AC43LO_2006A" |
+                         HUGGO$manyID == "NIC-VEN[LRD]_2007A" | HUGGO$manyID == "AC33LJ_2007A"|
+                         HUGGO$manyID == "ARG-ECU[LHE]_2007A" | HUGGO$manyID == "AC11PB_2007A" |
+                         HUGGO$manyID == "AD33LN_2007A" | HUGGO$manyID == "AC34LD_2007A" |
+                         HUGGO$manyID == "AD25LF_2008A" | HUGGO$manyID == "BRA-CRI[DSL]_2008A"|
+                         HUGGO$manyID == "BRA-CRI[PDB]_2008A" | HUGGO$manyID == "BRA-CRI[DCE]_2008P" |
+                         HUGGO$manyID == "AE22LE_2008A" | HUGGO$manyID == "AS35LO_2008A" |
+                         HUGGO$manyID == "CP09LA_2012O"| HUGGO$manyID == "CC09DM_1954A" |
+                         HUGGO$manyID == "VB13GV_2008O" | HUGGO$manyID == "CS18PS_1954A"|
+                         HUGGO$manyID == "CS18PS_1954A"),]
+
+# Stage five: Recode "9999-XX-XX-" Force date for PF07BS_1897A, RM04LF_1937A,
+# IP07TS_1971E, CPRAEM_2005P:CPRAEM_1999A, EF04GG_2007A
+# since this coding is no longer used for Force dates in HUGGO dataset.
+HUGGO <- HUGGO %>%
+  dplyr::mutate(Force = ifelse(grepl("9999-XX-XX-", Force), NA, Force))
+
+# Stage six: format data correctly for exporting
+HUGGO <- HUGGO %>%
+  dplyr::mutate(across(everything(),
+                       ~stringr::str_replace_all(., "^NA$", NA_character_))) %>%
+  dplyr::distinct() %>%
+  dplyr::mutate(Begin = messydates::as_messydate(Begin),
+                Signature = messydates::as_messydate(Signature),
+                Force = messydates::as_messydate(Force),
+                End = messydates::as_messydate(End)) %>%
+  dplyr::arrange(Begin)
+
+# Stage seven: Connecting data
 # Next run the following line to make HUGGO available
 # within the package.
-manypkgs::export_data(HUGGO, database = "agreements",
+manypkgs::export_data(HUGGO, datacube = "agreements",
                       URL = "Hand-coded data by the GGO team")
 # This function also does two additional things.
 # First, it creates a set of tests for this object to ensure
@@ -872,10 +796,3 @@ manypkgs::export_data(HUGGO, database = "agreements",
 # Second, it also creates a documentation file for you to fill in.
 # Please make sure that you cite any sources appropriately and
 # fill in as much detail about the variables etc as possible.
-
-# To reduce size of text data stored in package:
-# 1. after exporting HUGGO to agreements database,
-# load 'agreements.rda' in environment.
-# 2. Delete 'agreements.rda' in 'data' folder.
-# 3. Run `usethis::use_data(agreements, internal = F, overwrite = T, compress = "xz")`
-# to save compressed text data.