+++ tests/testthat/test_CIESIN.R
@@ -13,14 +13,16 @@ test_that("missing observations are reported correctly", {
 # Uniformity tests (agreements have a source ID, a string title, a signature and
 # entry into force date)
 test_that("datasets have the required variables", {
-  pointblank::expect_col_exists(agreements[["CIESIN"]], pointblank::vars(Title))
-  pointblank::expect_col_exists(agreements[["CIESIN"]], pointblank::vars(Beg))
+  pointblank::expect_col_exists(agreements[["CIESIN"]],
+                                pointblank::vars(Title))
+  pointblank::expect_col_exists(agreements[["CIESIN"]],
+                                pointblank::vars(Begin))
   expect_true(any(grepl("ID$", colnames(agreements[["CIESIN"]]))))
   pointblank::expect_col_exists(agreements[["CIESIN"]],
                                 pointblank::vars(Signature))
 })
 
-# Date columns should be in messydt class
+# Date columns should be in mdate class
 test_that("Columns are not in date, POSIXct or POSIXlt class", {
   expect_false(any(lubridate::is.Date(agreements[["CIESIN"]])))
   expect_false(any(lubridate::is.POSIXct(agreements[["CIESIN"]])))
@@ -28,17 +30,17 @@ test_that("Columns are not in date, POSIXct or POSIXlt class", {
 })
 
 # Dates are standardized for mandatory column
-test_that("Column `Beg` has standardised dates", {
-  expect_equal(class(agreements[["CIESIN"]]$Beg), "mdate")
-  expect_false(any(grepl("/", agreements[["CIESIN"]]$Beg)))
+test_that("Column `Begin` has standardised dates", {
+  expect_equal(class(agreements[["CIESIN"]]$Begin), "mdate")
+  expect_false(any(grepl("/", agreements[["CIESIN"]]$Begin)))
   expect_false(any(grepl("^[:alpha:]$",
-                         agreements[["CIESIN"]]$Beg)))
+                         agreements[["CIESIN"]]$Begin)))
   expect_false(any(grepl("^[:digit:]{2}$",
-                         agreements[["CIESIN"]]$Beg)))
+                         agreements[["CIESIN"]]$Begin)))
   expect_false(any(grepl("^[:digit:]{3}$",
-                         agreements[["CIESIN"]]$Beg)))
+                         agreements[["CIESIN"]]$Begin)))
   expect_false(any(grepl("^[:digit:]{1}$",
-                         agreements[["CIESIN"]]$Beg)))
+                         agreements[["CIESIN"]]$Begin)))
 })
 
 test_that("Column `Signature` has standardised dates", {
@@ -54,12 +56,12 @@ test_that("Column `Signature` has standardised dates", {
                          agreements[["CIESIN"]]$Signature)))
 })
 
-# Dataset should be ordered according to the "Beg" column
+# Dataset should be ordered according to the "Begin" column
 test_that("dataset is arranged by date variable", {
-  expect_true(agreements[["CIESIN"]]$Beg[1] <
-                agreements[["CIESIN"]]$Beg[10])
-  expect_true(agreements[["CIESIN"]]$Beg[50] <
-                agreements[["CIESIN"]]$Beg[75])
-  expect_true(agreements[["CIESIN"]]$Beg[100] <
-                agreements[["CIESIN"]]$Beg[120])
+  expect_true(agreements[["CIESIN"]]$Begin[1] <
+                agreements[["CIESIN"]]$Begin[10])
+  expect_true(agreements[["CIESIN"]]$Begin[50] <
+                agreements[["CIESIN"]]$Begin[75])
+  expect_true(agreements[["CIESIN"]]$Begin[100] <
+                agreements[["CIESIN"]]$Begin[120])
 })